Add page title to search modal
