+++ packages/guider/src/webpack/loader/md-loader.ts
@@ -1,5 +1,6 @@
 import { compile } from '@mdx-js/mdx';
 import remarkFrontmatter from 'remark-frontmatter';
+import type { Heading } from '@vcarl/remark-headings';
 import remarkHeadings from '@vcarl/remark-headings';
 import remarkHeadingId from 'remark-heading-id';
 import grayMatter from 'gray-matter';
@@ -8,10 +9,6 @@ import rehypeExtractExcerpt from 'rehype-extract-excerpt';
 import remarkLinkRewrite from 'remark-link-rewrite';
 import { remarkNpm2Yarn } from '@theguild/remark-npm2yarn';
 import remarkGfm from 'remark-gfm';
-import { SKIP, visit } from 'unist-util-visit';
-import type { Heading, HeadingData, Root } from 'mdast';
-import { phrasing } from 'mdast-util-phrasing';
-import type { VFileWithOutput } from 'unified';
 import {
   transformerNotationDiff,
   transformerNotationHighlight,
@@ -19,13 +16,9 @@ import {
   transformerNotationWordHighlight,
   transformerNotationErrorLevel,
 } from '@shikijs/transformers';
+import { remarkSearchData } from './search-data';
 
 const EXPORT_FOOTER = 'export default ';
-
-interface Section {
-  heading?: { id: string; depth: Heading['depth']; text: string };
-  content: string;
-}
 
 export async function mdLoader(source: string) {
   const meta = grayMatter(source);
@@ -75,67 +68,7 @@ export async function mdLoader(source: string) {
           },
         },
       ],
-      () => {
-        return (root: Root, vFile: VFileWithOutput<any>) => {
-          const sections: Section[] = [];
-
-          let currentSection: Section | undefined;
-          let previousParentNode: any;
-
-          visit(root, (node, _, parent) => {
-            if (node.type === 'heading') {
-              if (currentSection) {
-                sections.push(currentSection);
-              }
-
-              const heading = node;
-              const id = (heading.data as HeadingData & { id: string })?.id;
-              const depth = heading.depth;
-              let text = '';
-              visit(heading, ['text', 'inlineCode'], (hChild: any) => {
-                text += hChild.value;
-              });
-
-              currentSection = {
-                heading: {
-                  id,
-                  depth,
-                  text,
-                },
-                content: '',
-              };
-
-              return SKIP;
-            }
-
-            const types = ['text', 'inlineCode'];
-
-            if (types.includes(node.type)) {
-              if (!currentSection) {
-                currentSection = {
-                  heading: undefined,
-                  content: '',
-                };
-              }
-
-              if (
-                previousParentNode &&
-                previousParentNode !== parent &&
-                !phrasing(previousParentNode)
-              ) {
-                currentSection.content += ' ';
-              }
-              currentSection.content += (node as any).value;
-
-              previousParentNode = parent;
-            }
-          });
-
-          if (currentSection) sections.push(currentSection);
-
-          vFile.data = { ...vFile.data, sections };
-        };
-      },
+      remarkSearchData,
     ],
     rehypePlugins: [
       rehypeExtractExcerpt,
@@ -168,6 +101,10 @@ export async function mdLoader(source: string) {
     excerpt: file.data.excerpt,
   };
 
+  const firstHeading = (file.data.headings as Heading[]).find(
+    (h) => h.depth === 1,
+  );
+
   const script = `
     import { createMdxPage } from "@neato/guider/client";
 
@@ -185,6 +122,7 @@ export async function mdLoader(source: string) {
     script,
     searchData: {
       sections: file.data.sections,
+      pageTitle: meta.data?.title ?? firstHeading?.value ?? undefined,
     },
   };
 }
